// app/home/type.ts
export interface Link {
  createdAt: string;
  lastReadAt: string;
  linkBookId: string;
  linkBookName: string;
  linkId: string;
  readCount: number;
  tags: string[];
  thumbnailURL: string;
  title: string;
  updatedAt: string;
  url: string;
  userId: string;
}

export type Banner = {
  clickURL: string;
  id: string;
  imageURL: string;
};